refactor: share README utils
